Changes of version: 20100529

+++ libesedb/libesedb_page_tree_values.c
@@ -38,7 +38,7 @@
  */
 int libesedb_page_tree_values_initialize(
      libesedb_page_tree_values_t **page_tree_values,
-     int number_of_value_definitions,
+     uint32_t number_of_value_definitions,
      liberror_error_t **error )
 {
 	static char *function = "libesedb_page_tree_values_initialize";
@@ -91,7 +91,7 @@ int libesedb_page_tree_values_initialize(
 		}
 		if( libesedb_array_initialize(
 		     &( ( *page_tree_values )->value_definition_array ),
-		     number_of_value_definitions,
+		     (int) number_of_value_definitions,
 		     error ) != 1 )
 		{
 			liberror_error_set(
